ST40: Adapt input validator to new project structure
